성적 부여 생성 및 이수 자동 생성

+++ src/main/java/edu/du/campusflow/repository/LectureRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface LectureRepository extends JpaRepository<Lecture, Integer> {
+public interface LectureRepository extends JpaRepository<Lecture, Long> {
     @Query("SELECT l.lectureId FROM Lecture l WHERE l.member.memberId = :memberId")
     List<Long> findLectureIdsByMemberId(Long memberId);
 }
